Moved methods to register and save tables to context
